fix decimal type

+++ Models/Order.cs
@@ -3,8 +3,8 @@
     public class Order
     {
         public int Id { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Quantity { get; set; }
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
     }
 }
