+++ Models/PreviousOrders.cs
@@ -4,9 +4,9 @@ namespace CyberMall.Models
 {
     public class PreviousOrders
     {
-        public double TotalAmount { get; private set; }
-        public double TaxAmount { get; set; }
-        public double ShippingCost { get; set; }
+        public decimal TotalAmount { get; private set; }
+        public decimal TaxAmount { get; set; }
+        public decimal ShippingCost { get; set; }
         public DateTime PurchaseDate { get; set; }
     }
 }
